lexical-graph/src/graphrag_toolkit/lexical_graph/requirements.txt
anthropic-bedrock==0.8.0
boto3>=1.36.1
botocore>=1.36.1
json2xml==5.2.0
llama-index-core==0.13.1
llama-index-embeddings-bedrock==0.6.1
llama-index-llms-anthropic==0.8.3
llama-index-llms-bedrock-converse==0.8.2
lru-dict==1.3.0
pipe==2.2
python-dotenv==1.0.1
smart_open==7.1.0
spacy==3.7.5
tfidf_matcher==0.3.0